+++ runtime/src/commonTest/kotlin/pbandk/json/OutputDefaultValuesTest.kt
@@ -1,6 +1,7 @@
 package pbandk.json
 
 import kotlinx.serialization.json.*
+import pbandk.testpb.MessageWithEnum
 import pbandk.testpb.TestAllTypesProto3
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -8,6 +9,8 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 class OutputDefaultValuesTest {
+    private val jsonConfigCompactOutput = JsonConfig.DEFAULT.copy(compactOutput = true)
+
     @Test
     fun testOutputDefaultValues_false() {
         val testAllTypesProto3 = TestAllTypesProto3(optionalString = "")
@@ -29,12 +32,12 @@ class OutputDefaultValuesTest {
         assertEquals(JsonPrimitive(""), parsedJson["optionalBytes"])
         assertEquals(JsonPrimitive(TestAllTypesProto3.NestedEnum.fromValue(0).name!!), parsedJson["optionalNestedEnum"])
 
-        assertEquals(JsonNull, parsedJson["optionalNestedMessage"])
+        assertFalse("optionalNestedMessage" in parsedJson)
 
         assertEquals(JsonArray(emptyList()), parsedJson["repeatedString"])
         assertEquals(JsonObject(emptyMap()), parsedJson["mapBoolBool"])
 
-        assertEquals(JsonNull, parsedJson["optionalStringWrapper"])
+        assertFalse("optionalStringWrapper" in parsedJson)
         assertEquals(JsonPrimitive(false), parsedJson["optionalBoolWrapper"])
     }
 
@@ -47,5 +50,42 @@ class OutputDefaultValuesTest {
         assertEquals(JsonPrimitive(false), parsedJson["optionalBool"])
         assertEquals(JsonNull, parsedJson["optionalString"])
         assertEquals(JsonPrimitive(""), parsedJson["optionalStringWrapper"])
+    }
+
+    @Test
+    fun testProto2EnumUnsetField() {
+        val message = MessageWithEnum()
+        assertEquals("{}", message.encodeToJsonString(jsonConfigCompactOutput))
+    }
+
+    @Test
+    fun testProto2EnumFieldSetToDefaultValue() {
+        val message = MessageWithEnum(value = MessageWithEnum.EnumType.FOO)
+        assertEquals("{\"value\":\"FOO\"}", message.encodeToJsonString(jsonConfigCompactOutput))
+    }
+
+    @Test
+    fun testProto2EnumFieldSetToNonDefaultValue() {
+        val message = MessageWithEnum(value = MessageWithEnum.EnumType.BAR)
+        assertEquals("{\"value\":\"BAR\"}", message.encodeToJsonString(jsonConfigCompactOutput))
+    }
+
+    @Test
+    fun testProto3EnumUnsetField() {
+        // Note that this is equivalent to 'testProto3EnumFieldSetToDefaultValue', adding for completeness.
+        val message = TestAllTypesProto3()
+        assertEquals("{}", message.encodeToJsonString(jsonConfigCompactOutput))
+    }
+
+    @Test
+    fun testProto3EnumFieldSetToDefaultValue() {
+        val message = TestAllTypesProto3(optionalNestedEnum = TestAllTypesProto3.NestedEnum.FOO)
+        assertEquals("{}", message.encodeToJsonString(jsonConfigCompactOutput))
+    }
+
+    @Test
+    fun testProto3EnumFieldSetToNonDefaultValue() {
+        val message = TestAllTypesProto3(optionalNestedEnum = TestAllTypesProto3.NestedEnum.BAR)
+        assertEquals("{\"optionalNestedEnum\":\"BAR\"}", message.encodeToJsonString(jsonConfigCompactOutput))
     }
 }